+++ src/fairchem/core/models/equiformer_v2/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .equiformer_v2_oc20 import EquiformerV2 as EquiformerV2
+from .equiformer_v2 import EquiformerV2
 
 __all__ = ["EquiformerV2"]
